pyproject.toml
[tool.poetry]
name = "podcast-cutter"
version = "0.1.0"
description = ""
authors = ["ilyamed <0ilya@bk.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.1.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
